+++ test/prototype/blockwise_fp8_inference/test_blockwise.py
@@ -11,12 +11,8 @@ from packaging import version
 
 triton = pytest.importorskip("triton", reason="Triton required to run this test")
 
-from torchao.kernel.blockwise_quantization import (
-    blockwise_fp8_gemm,
-    fp8_blockwise_act_quant,
-    fp8_blockwise_weight_dequant,
-    fp8_blockwise_weight_quant,
-)
+from torchao.prototype.blockwise_fp8_inference.kernels import fp8_blockwise_act_quant
+from torchao.quantization.quantize_.workflows.float8.kernels import _blockwise_fp8_gemm
 from torchao.utils import is_sm_at_least_90
 
 BLOCKWISE_SIZE_MNK = [
@@ -29,18 +25,14 @@ BLOCKWISE_SIZE_MNK = [
 ]
 
 
-@pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
-@pytest.mark.skipif(not is_sm_at_least_90(), reason="Requires CUDA capability >= 9.0")
-@pytest.mark.parametrize("_, N, K", BLOCKWISE_SIZE_MNK)
-@pytest.mark.parametrize("dtype", [torch.float8_e4m3fn, torch.float8_e5m2])
-def test_blockwise_quant_dequant(_, N, K, dtype):
-    x = torch.randn(N, K).cuda()
-    qx, s = fp8_blockwise_weight_quant(x, dtype=dtype)
-    x_reconstructed = fp8_blockwise_weight_dequant(qx, s)
-    error = torch.linalg.vector_norm(x - x_reconstructed) / torch.linalg.vector_norm(x)
-    print(f"Relative Error: {error.item():.6f}")
-
-    assert error < 0.1, "Quant-Dequant error is too high"
+def _weight_quant_reference(w: torch.Tensor, block_size: int, dtype: torch.dtype):
+    # Eager reference for block_size x block_size blockwise fp8 weight quant,
+    # matching the semantics used by _blockwise_fp8_gemm's b/b_scale operands.
+    n, k = w.shape
+    wr = w.reshape(n // block_size, block_size, k // block_size, block_size)
+    s = wr.abs().amax(dim=(1, 3), keepdim=True).float() / 448.0
+    w_q = (wr / s).to(dtype).reshape(n, k)
+    return w_q, s.reshape(n // block_size, k // block_size)
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA not available")
@@ -56,8 +48,8 @@ def test_blockwise_fp8_gemm(M, N, K, dtype):
     B = torch.randn(N, K).cuda()
     C = A @ B.T
     A_q, A_s = fp8_blockwise_act_quant(A, dtype=dtype)
-    B_q, B_s = fp8_blockwise_weight_quant(B, dtype=dtype)
-    C_q = blockwise_fp8_gemm(A_q, A_s, B_q, B_s)
+    B_q, B_s = _weight_quant_reference(B, block_size=128, dtype=dtype)
+    C_q = _blockwise_fp8_gemm(A_q, A_s, B_q, B_s)
     assert C_q.dtype == torch.bfloat16, "unsupported"
     error = torch.linalg.vector_norm(C - C_q) / torch.linalg.vector_norm(C)
     print(f"Relative Error: {error.item():.6f}")
