+++ torchao/prototype/blockwise_fp8_inference/__init__.py
@@ -1,16 +1,7 @@
-from torchao.kernel.blockwise_quantization import (
-    blockwise_fp8_gemm,
-    fp8_blockwise_act_quant,
-    fp8_blockwise_weight_dequant,
-    fp8_blockwise_weight_quant,
-)
-
 from .blockwise_linear import BlockwiseQuantLinear
+from .kernels import fp8_blockwise_act_quant
 
 __all__ = [
-    "blockwise_fp8_gemm",
     "BlockwiseQuantLinear",
     "fp8_blockwise_act_quant",
-    "fp8_blockwise_weight_quant",
-    "fp8_blockwise_weight_dequant",
 ]
