+++ torchao/prototype/quantization/float8_static_quant/prototype_float8_tensor.py
@@ -22,9 +22,6 @@ from torchao.float8.inference import (
     preprocess_data,
     preprocess_scale,
 )
-from torchao.kernel.blockwise_quantization import (
-    blockwise_fp8_gemm,
-)
 from torchao.quantization.granularity import PerRow, PerTensor
 from torchao.quantization.quant_primitives import (
     _choose_scale_float8,
@@ -37,6 +34,9 @@ from torchao.quantization.quantize_.common import (
 )
 from torchao.quantization.quantize_.workflows import (
     QuantizeTensorToFloat8Kwargs,
+)
+from torchao.quantization.quantize_.workflows.float8.kernels import (
+    _blockwise_fp8_gemm,
 )
 from torchao.quantization.utils import get_block_size
 from torchao.utils import (
@@ -428,7 +428,7 @@ def _float8_addmm_impl(
                 # TODO(future PR): add mslk path if available
                 # TODO(future PR): proper out_dtype handling
                 assert _is_1_128_scaled(input_tensor), "unsupported"
-                res = blockwise_fp8_gemm(
+                res = _blockwise_fp8_gemm(
                     inpt_data,
                     input_scale,
                     w_data.t(),
